requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bidict==0.22.1
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
python-engineio==4.3.4
python-socketio==5.7.2
pyxel==1.9.11
yarl==1.8.2
